+++ cli/lib/plunder.ts
@@ -13,6 +13,7 @@ import { linkCheckingCommand } from './links.ts'
 
 const knownOpts: Record<string, any> = {
     all: Boolean,
+    ['booty-port']: [Boolean, Number],
     ['capture-hook']: String,
     ['confirm-install']: Boolean,
     ['css-breakpoints']: Boolean,
@@ -79,6 +80,16 @@ const mode = (function resolveMode(): PlunderMode | null {
     }
 })()
 
+function bootyPort(): false | number {
+    if (typeof parsed['booty-port'] === 'number') {
+        return parsed['booty-port']
+    } else if (typeof parsed['booty-port'] === 'undefined') {
+        return 7979
+    } else {
+        return !!parsed['booty-port'] ? 7979 : false
+    }
+}
+
 function isSkipConfirmInstall(): boolean {
     return parsed['confirm-install'] === true
 }
@@ -115,7 +126,10 @@ if (parsed.help) {
             recursive: parsed['recursive'] === true,
             urls: parsed.argv.remain.filter(remain => remain !== 'capture'),
         },
-        isSkipConfirmInstall(),
+        {
+            installConfirmed: isSkipConfirmInstall(),
+            bootyPort: bootyPort(),
+        },
     )
 }
 
@@ -168,6 +182,7 @@ Use a webapp UI to command screenshot capturing.
  ${ansi.bold('Options:')}
     ${ansi.bold('-b')}, ${ansi.bold('--browser')} <BROWSER>    Browser engine used when not specified by device emulation
                                [values: chromium (default) | firefox | webkit]
+        ${ansi.bold('--[no]-booty-port')}      Port to open HTTP server for web report [default: 7979]
         ${ansi.bold('--capture-hook')}         Path to a script ran before each screenshot capture
         ${ansi.bold('--confirm-install')}      Confirm Playwright browser installs
         ${ansi.bold('--css-breakpoints')}      Parse CSS and capture screenshots on media query breakpoints